% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adt_cci_analysis}
\alias{adt_cci_analysis}
\title{Infer the protein-protein celluar communication}
\usage{
adt_cci_analysis(
  object,
  input_list = NULL,
  id = NULL,
  lr = NULL,
  celltype_sender = NA,
  celltype_receiver = NA,
  celltype_pair = NULL,
  lr_pair = NULL,
  module_sender = NA,
  module_receiver = NA,
  mode = "1",
  transfer = F
)
}
\arguments{
\item{object}{Input data in SeuratObject format}

\item{input_list}{Spatrio module analysis results}

\item{id}{Dataframe for id transfer}

\item{lr}{Dataframe of cell-cell communication information}

\item{transfer}{}
}
\description{
Infer the protein-protein celluar communication
}
