% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatrio_resonance.R
\name{cor_filtering}
\alias{cor_filtering}
\title{Filter the genes with low correlation}
\usage{
cor_filtering(mat, min_cutoff = 0.5, min = 5, max = 100, na_diag = F)
}
\arguments{
\item{na_diag}{}
}
\description{
Filter the genes with low correlation
}
