% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gene_cci_analysis}
\alias{gene_cci_analysis}
\title{Infer the gene-gene celluar communication by SpaTrio}
\usage{
gene_cci_analysis(
  object,
  assay = "RNA",
  xcoord = NULL,
  ycoord = NULL,
  species = "Human",
  celltype = NULL,
  lrpairs = NULL,
  pathways = NULL,
  ...
)
}
\arguments{
\item{object}{Input data in SeuratObject format}

\item{assay}{Name of assay to use}

\item{xcoord}{x coordinates of all cells}

\item{ycoord}{y coordinates of all cells}

\item{species}{A character indicating the species of the spatial transcriptomics data. Either ‘Human’ or ‘Mouse’}

\item{celltype}{A character specifying the cell type of ST data. To bypass the deconvolution step and directly infer cell-cell communication, please provide the cell type. Default is NULL}
}
\description{
Infer the gene-gene celluar communication by SpaTrio
}
