% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gene_cci_plot}
\alias{gene_cci_plot}
\title{Visualization of gene-gene communication}
\usage{
gene_cci_plot(
  object,
  celltype_sender = NULL,
  celltype_receiver = NULL,
  top_lrpairs = 20,
  color = "#337EDE",
  return_data = FALSE,
  lr_pair = NULL,
  celltype_pair = NULL
)
}
\arguments{
\item{object}{Input data in SpaTalk format}

\item{celltype_pair}{}
}
\description{
Visualization of gene-gene communication
}
