% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gene_motif_lineplot}
\alias{gene_motif_lineplot}
\title{Visualization of selected gene-motif combos for celltypes}
\usage{
gene_motif_lineplot(object, feature_gene = NULL, motif_gene_df)
}
\arguments{
\item{object}{Input data in SeuratObject format}

\item{feature_gene}{Gene name}

\item{motif_gene_df}{Dataframe of motifs and genes}
}
\description{
Visualization of selected gene-motif combos for celltypes
}
