% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gene_motif_plot}
\alias{gene_motif_plot}
\title{Visualization of significant gene-motif combos for celltypes}
\usage{
gene_motif_plot(df.sig, plot = "sig", return_data = FALSE, col = NULL)
}
\arguments{
\item{df.sig}{Dataframe of significant gene-motif combos for celltypes}

\item{plot}{Select the type of gene-motif combos to plot. "sig","pos", or "neg"}

\item{return_data}{Whether to return data}

\item{col}{Color}
}
\description{
Visualization of significant gene-motif combos for celltypes
}
