% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sdplot}
\alias{sdplot}
\title{Visualization of spatial clustering data}
\usage{
sdplot(
  object,
  pt.size.factor = 1,
  image.alpha = 0,
  stroke = NA,
  highlight.clsuter = NULL,
  highlight.cell = NULL,
  shape = 21,
  subset = NULL,
  ident = NA,
  bg = "white",
  cols.highlight = c("#DE2D26", "grey80"),
  label.color = "white",
  ...
)
}
\arguments{
\item{object}{Input data in SeuratObject format}

\item{pt.size.factor}{Scale the size of the spots}

\item{image.alpha}{Transparency of background image}

\item{stroke}{The width of the border around the spots}

\item{highlight.clsuter}{The cluster name of cells to highlight}

\item{highlight.cell}{The cells to highlight}

\item{shape}{Adjust the shape of the spots}

\item{subset}{The cluster name of cells to splot}

\item{bg}{The color to fill the background}

\item{cols.highlight}{Colors to highlight the cells}

\item{label.color}{Color of the label text}

\item{...}{Extra parameters used in sdplot. Parameters passed to \code{\link[Seurat]{SpatialDimPlot}}}
}
\description{
Visualization of spatial clustering data
}
