% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sfplot}
\alias{sfplot}
\title{Visualization of spatial expression data}
\usage{
sfplot(
  object,
  stroke = NA,
  features = NA,
  shape = 21,
  option = "A",
  ncol = NULL,
  pt.size.factor = 1,
  ...
)
}
\arguments{
\item{object}{Input data in SeuratObject format}

\item{stroke}{The width of the border around the spots}

\item{features}{Features to plot}

\item{shape}{Adjust the shape of the spots}

\item{option}{A character string of color map option.}

\item{ncol}{Number of columns}

\item{pt.size.factor}{Scale the size of the spots}

\item{...}{Extra parameters used in sdplot. Parameters passed to \code{\link[Seurat]{SpatialFeaturePlot}}}
}
\description{
Visualization of spatial expression data
}
