% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatrio_resonance.R
\name{smoothScoresNN}
\alias{smoothScoresNN}
\title{Smooth feature matrix using kNNs}
\usage{
smoothScoresNN(nn_matrix, mat, geneList = NULL, subset = NULL, nCores = 1)
}
\arguments{
\item{nn_matrix}{Matrix of K Nearest neighbor indices for each cell (row), where each column is one of K nearest neighbors per cell.}

\item{mat}{Matrix of gene/adt expression or motif activity score. Column names must match the rownames of nn_matrix.}

\item{nCores}{Integer specifying the number of cores to use. Default is 1.}
}
\value{
A matrix of scores for each feature and each cell, smoothed over its K nearest neighbors.
}
\description{
Method to smooth sparse scores (e.g. RNA expression, ADT expression, or motif scores) per cell per feature using cell K nearest neighbors (NNs).
}
