% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{spatrio_boxplot}
\alias{spatrio_boxplot}
\title{Visualization of spatrio results using dotplot}
\usage{
spatrio_boxplot(
  df,
  x = NA,
  y = NA,
  fill = NA,
  xlab = "Cluster",
  ylab = NULL,
  compare = NULL,
  method = "wilcox.test",
  label = "p.signif",
  vjust = 0.2,
  cols = NULL,
  title = NULL,
  x_level = NULL
)
}
\arguments{
\item{df}{Dataframe collated from the results of spatrio}

\item{x}{Dataframe column name used as x axis data}

\item{y}{Dataframe column name used as y axis data}

\item{fill}{Dataframe column name used to color box}

\item{xlab}{Label of x axis}

\item{ylab}{Label of y axis}

\item{compare}{List of group for comparision}

\item{method}{A character string indicating which method to be used for comaprision}

\item{label}{A character string specifying label type of comaparision}

\item{vjust}{Vertical alignment adjustment of text}

\item{cols}{Group color}

\item{title}{Title of plot}
}
\description{
Visualization of spatrio results using dotplot
}
