% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatrio_resonance.R
\name{spatrio_resonance}
\alias{spatrio_resonance}
\title{Space resonance feature module}
\usage{
spatrio_resonance(
  object,
  sigma = NULL,
  assay = "RNA",
  feature_select = NULL,
  correlation = "pearson",
  maxK = 8,
  k = 8,
  min_avg_con = 0.5,
  min_avg_cor = 0.5,
  min_featuer = 20,
  max_featuer = 100,
  min_pct_cutoff = NULL,
  smooth = NULL,
  smooth_k = 10,
  reduction = NULL,
  scale = T,
  loc_mat = NULL
)
}
\arguments{
\item{object}{Input data in SeuratObject format}

\item{assay}{Which assay to use}

\item{min_pct_cutoff}{Filter features with low expression ratio}

\item{reduction}{Which low-dimensional representation to use for smoothing}

\item{loc_mat}{Coord matrix}

\item{...}{}
}
\description{
Space resonance feature module
}
